Cleaning up and breaking up into separate files just a bit
